Migrated connector from 3.5.3 to 3.6.0 and created Connection-Strategy for multiple modes of authentication.
For implementing the Connection-Strategy and Connection-Management component,
migrating the connector to 3.6.0 is required.

Tested the connector successfully using a mavenized flow, which in 3.5.3 used to
fail. However, unpredictable results with a non-mavenized flow: sometimes it
would work, sometimes it won't.

This commit involves a lot of critical changes that bundles up migrating the
connector to a new platform and implementing a new model of authentication. But
since they have close dependency, its just one commit with a lot of changes.